+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Add: options for SO test page:
   - Plasma limit and Plasma time;
   - Use Serial ports.
+- Add: Button to copy SO level options to clipboard to use in level upload
+  requests.
 - Add: "Remove Other files" button.
 - Add: "Close Other levels" button.
 - Enh: Reduce paddings in several UI components to compress whole UI.